examples/aria-service-proxy-plugin/requirements.txt
Jinja2==2.9
MarkupSafe==1.0
bottle==0.12.7
certifi==2017.11.05
chardet==3.0.4
cloudify-plugins-common==4.2
cloudify-rest-client==4.2
decorator==4.1.2
idna==2.6
networkx==2.0
pika==0.9.14
proxy-tools==0.1.0
requests==2.18.4
requests-toolbelt==0.8.0
urllib3==1.22
wsgiref==0.1.2
apache-ariatosca==0.2.0
